requirements.txt
streamlit
chromadb
google-generativeai
sentence-transformers
langchain
langchain-google-genai
python-dotenv
pandas
numpy
requests
beautifulsoup4
selenium
webdriver_manager
nltk
tqdm
pillow
jsonschema
pytest